third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-navigation-parent-expected.txt

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'sandboxed-iframe-navigation-source.html'. The frame attempting navigation is sandboxed, and is therefore disallowed from navigating its ancestors.

CONSOLE ERROR: Uncaught SecurityError: Failed to execute 'replace' on 'Location': The current window does not have permission to navigate the target frame to 'sandboxed-iframe-navigated.html'.
This test verifies that a sandboxed IFrame cannot navigate an ancestor in the frame tree by assigning the location attribute.

This is done by loading ten non-sandboxed IFrames, and a single sandboxed one. Expect ten frames to be navigated, but the sandboxed one to not be one of them.

This test will print "PASS" on success.

          

PASS
